New Anticursor class with ancestors map
Point::init should not create the ancestors map. Not only is the
ancestors map on a cursor unused, it's not kept up-to-date and quickly
becomes wrong. We shouldn't have a property of all Points that's unused
and wrong on some of them, and only correct and used on some, there
should just be a subclass of Points where that property will always be
correct.

This also presents the perfect spot to have a comment clearly explaining
what the heck the oddly-named "anticursor" is.

Point.copy() really didn't seem worth duplicating, it only has one
non-anticursor usage, I just got rid of it.