Replace .APIClass with .KIND_OF_MQ and API registry
Last bit of prep for a single controller to be accessed via API objects
of different interface versions. See, in the next commit,
`MathQuill.getInterface()` will create a whole new instance of the
entire API under the specified interface version:

    var MQ1 = MathQuill.getInterface(1);
    var MQ2 = MathQuill.getInterface(2);
    // MQ1 and MQ2 are two wholly separate instances of the entire API,
    // with different interface versions

    var mq1 = MQ1.MathField(mathFieldSpan);
    var mq2 = MQ2(mathFieldSpan);
    // mq1 and mq2 are API objects for the same math field, but they
    // have different interface versions

After `MQ2()` gets the controller for the math field, how does it know
to create a math field API object and not a static math or text field
API object? We can't use `.APIClass`, which would be an interface
version 1 API class, so we record a `.KIND_OF_MQ` string constant which
`MQ2()` can look up in the global registry of API classes, `API`.

Incidentally the API functions like `MQ.MathField()` exactly correspond
to the API classes in this API registry, so now instead of directly
setting API functions on `MQ` by creating them from API classes with
`APIFnFor()`, we can loop over the registry and know the name of the API
class to use for `.KIND_OF_MQ`.