getInterface() creates new copies of entire API
(You really want -w to view the diff.)

The first commit in this PR changed the call pattern from
`MathQuill.interfaceVersion(1)` to `MQ = MathQuill.getInterface(1)`, but
obviously didn't make it possible to actually support multiple interface
versions.

After commit after commit of prep work, it finally all comes together:
every call to getInterface() creates a whole new copy of the entire API,
closing over the specified interface version. The next commit can then
add interface version 2 and actually create separate copies of the API
with different behavior.

With the API trapped in the closure of getInterface(), each copy of the
entire API needs its own copies of all the API classes (we wouldn't want
an API object to be `instanceof` an API class of a different interface
version), so now the `API` registry has factory functions to create API
classes rather than API classes themselves. For the API classes to
subclass the right copy of superclass, the superclasses are now defined
on an `APIClasses` registry that's passed to the factory, which is also
the registry on which `MQ()` can look up the right copy of an API class
given `.KIND_OF_MQ`.

Same old 2 complications:
- `MathQuillMathField::registerInnerField()` is independent of interface
  version, so pass in the right copy of the math field API class
- `Controller::handle()` is also independent of interface version, but
  handlers ought be called with the interface version they were defined
  with, so record that alongside the handlers object
    + this has to be a special case in config() because optionProcessors
      don't have access to `APIClasses` and it's not obvious they should

Final note, we can now explicitly get a copy of the API at interface
version 1 and use it for the pre-interface-versioning shim.