ARIA alert modification:
NVDA is the only AT which spits out the word 'alert' any time an ARIA alert fires for something given the aria role of, well, alert. Remove this attribute and depend on remaining attrs to cause AT to speak.

Works with ZoomText Fusion, Window-Eyes, and NVDA. VoiceOver needs to be tested.