More ARIA speech clean-up
Moving to beginning and end of blocks with the Arrow keys now causes MQ to speak the containing element, but Arrowing in something with an existing ARIA label (e.g. numerator/denominator) doesn't cause the element name to be repeated with each keypress unless block boundary is reached.

In cases where the container is just a nondescript block, speak 'block' when reaching beginning or end of bounds.