Fix #584: Hide iOS blinking blue cursor WITHOUT breaking typing
In Safari on my iPhone 5 running iOS 9.2.1, the blinking blue cursor
doesn't show up at all when focusing and typing in this thing:
    http://jsbin.com/sodaji/edit?html,css,js,output
And in all browsers typing in that works just fine, including typing
newlines by pressing Enter or combos like Shift-Enter.

And with this patch, typing on /test/demo.html works great on my iPhone:
    ![screenshot](https://git.io/vaoNb)
(Notice that it's not zoomed wayyyy in, like it used to)

As detailed by #584:
Fixes #540
Fixes #566
Hopefully addresses #559?