Initialise cursor in StaticMath blocks
We were experiencing an issue where Mathquill fields
were not flowing correctly.  The width of the element
was not calculated correctly, so in Chrome they would
sometimes render in vertically offset locations, as it tried
to reconcile the flow.  Bizarrely, clicking on the StaticMath
block would cause the Mathquill to snap back into place.

It turned out this was happening because even unfocused
StaticMath blocks had a virtual cursor created and
rendered.

Haydn and I read through the code to try to figure out why
the cursor is being rendered at all in StaticMath blocks.
It seems that the code replaces the focus blur event
handlers with just the initialisation code from that handler,
but omits the line which initializes the cursor, so it gets
rendered into the HTML in a weird state.

Here we just include the extra initialisation line which
resolves our issues.