Add support for the set of standard amsmath LaTeX commands
These are mostly additive changes (i.e. adding a definition for a
command that did not exist before), but there are a few notable
exceptions listed below.

A remaining issue is that there is a larger set of delimiters recognized
as "Bracket". However, delimiters that are "Bracket" can't seem to
function independently. That is, "\lceil x \rceil" does not render, but
"\left\lceil x \right\rceil" does.​

\deg was aliased to \degree *AND* defined as a built-in function. \deg
*is* a standard amsmath commands, therefore dropped the alias

\ast was aliased to \star, but they are different standard amsmath
commands

\asymp was aliased to \approx, but they are different standard amsmath
commands

\diamond, \bigtriangleup, \ominus, \uplus, \bigtriangledown, \sqcap,
\triangleleft, \sqcup, \triangleright, \odot, \circledot, \bigcirc,
\dagger, \ddager, \wr and \amalg were defined as VanillaSymbol, but they
are BinaryOperators.

\models, \prec, \succ, \preceq, \succeq, \simeq, \mid, \ll, \gg,
\parallel, \nparallel, \bowtie, \sqsubset, \sqsupset, \smile,
sqsubseteq, \sqsupseteq, \doteq, \frown, \vdash, \dashv, \nless, \ngtr
were defined as VanillaSymbols, but they are BinaryOperators (comparison
operators)

arrows and harpoons were defined as VanillSymbol but they are
BinaryOperatos
\oint, \bigcap, \bigcup, \bigsqcup, \bigvee, \bigwedge, \bigodot,
\bigotimes, \bitoplus, \biguplus were defined as VanillaSymbol, but they
are variable size operators (SumationNotation)

\implies was aliased to \Rightarrow, but it should be to \Longrightarrow

\iff was aliased to \Leftrightarrow, but it should be to
\Longleftrightarrow

\oslash was aliased to \varnothing, but there is a standard amsmath
command, therefore dropped this alias

\Upsilon was mapped to &upsih; U+03d2 (GREEK UPSILON WITH HOOK SYMBOL).
It should be mapped to U+03a5 (GREEK CAPITAL LETTER UPSILON). \upsih and
\Upsih now maps to &upsih;

\wedge was mapped to \land. They are distinct commands.

\vee was mapped to \lor. They are distinct commands.