Rename Parser::{parse => _parse}
Prep for custom symbols, which are gonna have a special contract related
to ENV_VARs. We want to allow all the parser combinator methods to be
called by anyone, but restrict parsing to latex.js

While we're at it, simplify Parser::_parse's usage for the particular
way we're using it, even though it's a totally weird contract (which is
now more ok because its usage is restricted to one place, latex.js), and
comment about that.

Also, fix text.test.js not to use latexMathParser directly, instead
initialize a controller and stuff and use that to test.