Strawman: Custom symbols option via ENV_VAR
ENV_VARs are explained in big comment where the system is set up.
The important bit for us is that we can pass in the dictionary of
backslash commands that we accept, including custom symbols, to the
`controlSequence` parser, without having to pass it through all
intervening function calls, using the ENV_VAR `localLatexCmds`.

That's exactly what `latexMathParser.parse()` now does with its new
required second argument, the dictionary of backslash commands to
accept. (Note that `LatexFragment`s are made of builtins, so they pass
in the dictionary of builtins to the parser.)

One quirk is that our options processing architecture doesn't let you
remap names. We want the API consumer to pass in `customSymbols`, and we
want a combined dictionary of the built-in backslash commands plus the
custom ones for the LaTeX parser, so the most natural thing to do is for
the `customSymbols` option processor to combine the supplied custom
symbols with the builtins into a combined dictionary. Because of said
quirk, this means that `controller.options.customSymbols` actually ends
up with the combined dictionary, not just custom symbols.

In addition to the LaTeX parser, `LatexCommandInput::renderCommand()`
also uses the combined dictionary.

Another quirk is that to avoid relying on `Objects.create()`, we use Pjs
to set up the prototype relationship from the custom+builtin combined
dictionaries to the builtins (that's `mkLatexCmds()`).

Open Question: Should custom symbols be allowed to override builtins?
It seems like something that should fail by default, but maybe like:

    MQ.config({
      customSymbols: {
        // maybe a flag to intentionally override builtins
        $overrideBuiltins: true,
        // or maybe even name the specific builtins
        $overrideBuiltins: 'approx asymp cong',
        // or just on a per-symbol basis
        approx: { ch: '≃', kind: 'rel', overrideBuiltin: true }
      }
    });

Final Note: LOL need <meta charset=...> on visual test page to be able
to use Unicode in the source.