Faster MathJax loading in reference.html
Switched to using a custom configuration of MathJax and the <script> tag
instead of inline $$. This results in much faster rendering by MathJax.
After correctly disabling initial typesetting I tried doing a manual
typesetting on startup but this resulted in **much** slower rendering
(because each queued typesetting operation results in a
relayout/repaint, which is very, very slow).