Make `target` an optional argument to .clickAt()
It previously was the first argument but was unused, which, why.

It's optional because it's not fundamentally necessary to the
functionality, but it's nice to allow it as a micro-optimization:
it's one less DOM FFI call, and it'll be available in pretty much any
use case I can think of (touch tap, drag-and-drop a thing into
MathQuill).