Revert some unicode character LatexCmds
These changes slipped in at

https://github.com/sclower/mathquill/commit/f53cfb10caacbc1f9bf3b1d31f5589cabb78a199

I don't fully understand how this problem came about. @sclower
explained to me that there was an issue with an encoding
byte-order-mark that was preventing minification from running properly
(on windows), so he updated the file to not have that byte order mark.

Maybe things got garbled during that translation? It sounds like one of
our other contributors might have an editor configured in a way that is
adding a BOM that windows can't handle, but I'm not sure who/when that
happened. I'd love to have a better understanding of what's going on
here so that we don't run into this again.