Add screenshot.js file to scripts
This script is used to connect to Selenium on Sauce using `wd` (a
webdriver for Selenium). This was the first webdriver that we came
across it seems to be feature complete, but the standard webdriver for
Selenium is http://webdriver.io/. The variable `browserVersions is the list of browsers that screenshots will be taken for.

This script assumes the following:
  1. You've installed wd with `npm install wd'.
  2. You've set the environment variables $SAUCE_USERNAME and $SAUCE_ACCESS_KEY.
  3. If the environment variable $CIRCLE_ARTIFACTS is not set imanges will be saved in /tmp

The intention of this script is that it will be ran from CircleCI

Example usage:
node screenshots.js http://localhost:9292/test/visual.html
node screenshots.js http://google.com