Support taking screenshots on Safari
Previously Sauce couldn't connect to the test server. After a bit of
research we discovered that Safari and Sauce don't work together unless
the server that Sauce is connecting to is ran on port 3000 or 4000.
https://david263a.wordpress.com/2015/04/18/fixing-safari-cant-connect-to-localhost-issue-when-using-sauce-labs-connect-tunnel/

Once Sauce could connect to the browser we also hit an issue where the
css wasn't loading in Safari. To get around this we set the
`Content-Type` header appropriately in the test server.