Fix sup/subscript tests of typing '³'
LatexFragment doesn't have .ctrlSeq, so .ctrlSeq.search() is erroring.
Even after fixing that to do === or something instead of .search(),
though, the subsequent .mathspeak() call (in aria.queue()) still errors,
because LatexFragment isn't a normal node that actually gets inserted
into the tree.

Instead, refactor so that rather than special-casing Fraction, we use
polymorphism: we extend the contract of Node::mathspeak() to take an
optional `createdLeftOf` parameter that will be the cursor if currently
creating the command. (In the next commit, this will actually be called
by Node::createLeftOf().)

Fraction::mathspeak() then knows in that case to speak the block that
the cursor is in.

And, LatexFragment::mathspeak() actually works, unlike the overridden
MathCommand::mathspeak() which breaks on LatexFragment since it's not a
normal node.