Refactor ARIA out of ctrlr.typedText() to block.write()
This feels good in part because there are no other ARIA calls in
services/textarea.js, but lots of others in commands/math.js, so this
feels like the right layer. Also, it was weird to .chToCmd() a temporary
command object to get the mathspeak, when there's an actual command
being created.

Unfortunately this diff is a little complex, because while there's only
one Controller::typedText() definition and two call sites (DOM event and
public API), there are several calls to .write() and even more
definitions that could be called. Stupid polymorphism.

They've been identified with `git grep -F '.write(' src` and
`git grep -F '.write =' src`.

We'll start with call sites, there's fewer of those. The call sites are:
- Controller::typedText() (definitely want ARIA),
- the .write() public API method (probably want ARIA),
- LatexCommandInput::write() (typing in a LatexCommandInput, definitely
    want ARIA)
- TextBlock::createLeftOf() (want ARIA to announce the whole
    .replacedText, not just the last character; as far as I can tell,
    passing in the whole string instead of character-wise just works)

Great so all the callers are happy that .write() calls out to ARIA.

As for the definitions of .write():
- MathBlock::write() in commands/math.js: this is the common case
- LatexCommandInput::write(): this is an interesting one. We probably
    want to announce each letter of a control word as it's typed (unless
    that would happen automatically anyway if we didn't override it?);
    if a non-letter is typed, we want to announce _both_ the newly-
    rendered backslash command, _and_ the non-letter command, so we
    queue up the newly-rendered command and let it be announced by the
    aria.alert() called by the .write() (or aria.alert() ourselves if
    .write() is skipped).
- SupSub::write(): with certain options, typing certain things "breaks
    out of" the subscript; we probably want to announce the "breaking
    out" by saying "Baseline" or something, right?
- TextBlock::write() in commands/text.js should announce the character
    that was typed