saneKeyboardEvents handles cut/copy, not just paste
Requested by @mikehaverstock:
https://github.com/desmosinc/mathquill/pull/27#discussion_r65231110

It really should've in the first place, as the source-of-truth on
keyboard/textare events.

--

Previously listened for `copy` event on textarea for both editable
fields and static math, now only for static math (which doesn't use
`saneKeyboardEvents()`), for editable fields the `cut` and `copy`
handlers are just methods on the controller object.