Fix diacritics to not collide with right border
Before:  https://git.io/voRN6
After:   https://git.io/voRNM

For static math, it's a simple matter of adding a CSS rule for
`.mq-root-block > .mq-diacritic:last-child`. Unfortunately, for editable
math the cursor can be the last DOM element, and the diacritic can be
inside a selection, and we don't want the extra right padding for
diacritics to change as the cursor or selection moves around, so that
complicates the CSS selector a bit. Luckily, width is pretty easy to add
a JS unit test for.