CI: Name the logfiles of all backgrounded tasks
...by redirecting stdout and stderr to named logfiles at the outermost
level, instead of using CircleCI's inscrutable default names:
https://github.com/circleci/circleci-docs/commit/93893a57b836ccfe173cf130e87a09bc1dab19b4#diff-4d4c15e72df421fa0e9d59645ecdc0b3R35

Also, merge stderr into stdout so we see the same thing we'd see running
these commands by hand in the command-line. (With separate stdout and
stderr, we can't see the order in which lines were output.)

Except for the trivial Node server that listens for the Mocha XUnit test
report, continue to redirect stdout of that to $CIRCLE_TEST_REPORTS/,
but do it at the outer level so that CircleCI will know stdout is
redirected.

Recommend -w to view diff.