Close #563: Add "ultra-bare-bones" quickstart.html
In addition to being helpful to someone who just downloaded MathQuill
for the first time, it makes it easy to sanity test release tarballs as
packaged by `npm pack`.

Considered putting it under test/, but it makes more sense in the root
of a release package, and it's equally as helpful to someone who just
`git clone`d MathQuill for the first time, so I think it's fine in the
root of the repo.