Output Unicode char for &Prime;, not double-quote
Also, parse the Unicode char, so that it'll be symmetric.

Note that "real LaTeX" (pdfTeX) requires the utf8x package to work with
this Unicode char. It just works with MathJax and the newer TeX engines
XeTeX and LuaTeX, though, and I can't find any way to render that
character with pdfTeX out-of-the-box (in fact there's apparently no way
to render arbitrary Unicode characters by hex code [1]), so, this is the
best we got. Discussion:
https://github.com/mathquill/mathquill/pull/648#issuecomment-254374445

[1]: http://tex.stackexchange.com/questions/289982/how-to-type-unicode-in-plain-tex/290011#comment703254_290032