CI: Use CircleCI build # for Sauce build name, not commit
I realized that commit hash is a bad build identifier because there can
be more than one build for a given commit, for example if we request a
rebuild; whereas there's a one-to-one correspondence between Sauce Labs
unit test runs and CircleCI builds, so if we simply link back to the
CircleCI build, we can access all the rest of that information.

We could just do the CircleCI build number or URL, but having seen how
it shows up in the Sauce Labs dashboard, a human-readable name is really
nice, just like CircleCI's list of builds.