Wrap non-numeric vanilla symbol mathspeak in quotes
Apple voices in VoiceOver (such as Alex, Bruce, and Victoria) do
some strange pronunciation given certain expressions,
e.g. "y-2" is spoken as "ee minus 2" (as if the y is short).
Not an ideal work-around, but placing quotation marks around
non-numeric vanilla symbols works. This will be reported to Apple.