CI: Convert script/screenshots.js to "promise chain" style
...of `wd` (our Node.js client for webdriver)

There's fewer indents and callbacks, so it's more readable overall.
There's still lots of indents though because in order to use info from
previous steps, have to use .then() callbacks, promises don't eliminate
callbacks, just uniformize them and in particular, uniformize
error-handling:

Instead of all the `if (err) log(err);`s, can have just one .fail()
handler at the end of the chain. On the other hand, injecting logging
steps in the chain is more annoying, had to define a `willLog()` promise
to log when each step completes.

We also use the fancy-ish Q feature .spread() to parallelize getting
scrollHeight and clientHeight (which requires combining logging steps),
not sure if we should be doing more or less of that.

For convenience, we move assembling the screenshot filename a little
earlier, to the nearest .then(). Also, unfortunately, we have to
manually wrap Node's filesystem API in a promise; the builtin Promise
should be available in Node 4+.

Recommend -w to view diff.