Remove vestigial comment in advancedSymbols.js
TeX backslash control sequences are letters-only, whereas HTML character
entity references are alphanumeric; therefore `&there4;` is a valid
character entity reference, but `\there4` cannot be a valid control
sequence. Initially I hoped to support every character entity reference
as a backslash command, but we can't support this one, so forget it.

The out-of-place "case" is because backslash commands used to be a
switch/case, not a dictionary:
https://github.com/mathquill/mathquill/commit/77bd6c476f081979c6ea754ece77eff783f3c928#diff-239c238d8c090904fa1a9261bf2ef674L961