CI: Fix screenshots.js to get only browser/OS version from Sauce
The browser and OS names reported by Sauce are inconsistent (Chrome on
macOS is reported as being on "Mac OS X"; Firefox and Safari on macOS is
reported as being on "MAC") and poorly formatted (all-caps or all
lowercase).

Note that .platformVersion is only reported by some drivers, so it's
optional.