CI: Fix and simplify glob for baseline screenshot
...for evergreen browsers

Whoops previous commit was broken because the double quotes were
preventing the glob from expanding. So fix that but also I discovered
that we don't need the `ls` or the `sh -c` if we don't do the double
quotes, just having another layer of subshell (the `$(...)`) will expand
the glob.