CI: Consolidate CircleCI commands for readability
This way there are only 3 "commands" (really, entire scripts) that
CircleCI runs, each which begins with a descriptive comment.

This commit also reduces the number of backgrounded commands, and most
backgrounded commands start with a word in the comment that describes it
well, so the 'stdout_*.txt' and 'stderr_*.txt' log outputs auto-captured
by CircleCI will have more readable names and we don't need those *.log
files.

Recommend -w to view diff.