Add special case parsing for \%\operatorname{of}.
Without this change to the parser, parsing `\%\operatorname{of}` and
then asking for LaTeX produces `\%of` instead of `\%\operatorname{of}`.

Mathquill unfortunately doesn't do a great job of tracking which things
were wrapped in `\operatorname` in expressions that it parses. Instead,
it relies on the list of `autoOperatorNames` to do unitalicizing, and
to resynthesize `\operatorname` during serialization.

An alternative here is to add `of` as an autoOperatorName, but I don't
think we want to do that because I think there's too high a change that
someone has written an implicit multipliclation of the variables `o`
and `f` in the past, and I don't want to break graphs that did that.

One corner case here is that if someone pastes in an expression with a
percent sign in it, and then tries to write `of` after the percent
sign, it will stay `\%of` and won't be upgraded to
`\%\operatorname{of}`. To get correct behavior in the calculator,
they'll need to delete the `%` and type a new one to get the special
`percentof` node. It's possible we could fix this by adding
siblingAdded behavior to the `%` node, but I'm not confident about
doing the necessary surgery on the node tree. I think this will be
uncommon enough in practice that we can get away with ignoring it.