Fix bug with missing cursor in IE.
Tracked the issue down to the fact that cursor.show() was being called
earlier after a renderLatexMath call than it previously was, because of
the newly added 'edit' notification.

I'm not totally sure why this was causing an issue in IE but not in
other browsers, but we can work around the problem by avoiding calling
'show' for the first time in response to an 'edit' notification.