Reduce amount of work that child iteration does.
Creating a mathquill Fragment is expensive because it initializes a
jQuery datastructure to represent the fragment. Fragment also exposes
some useful iteration methods, and so mathquill creates a Fragment
whenever it needs to iterate over a group of nodes. This does a lot
more work than necessary in cases when the jQuery datastructure will
never be used: for example, serializing to latex, text, or mathspeak.

After this change, the time it takes to serialize a 20 digit number to
latex or mathspeak drops by almost an order of magnitude, from about
200 microseconds to about 30 microseconds.

The time it takes to serialize a 20-fold nested square root drops from
about 800 microseconds to about 80 microseconds.

The foldChildren/eachChild and fold/each methods of Node and Fragment
only depend on the `ends` datastructure that both of these classes
share, so they can share a fold/each implementation that takes the
`ends` datastructure as an argument.